Moved the reclassify link methods to a separate file
